src/arew/epoll: register events with a continuation.
Relying on a continuation make things cleaner in untangle.

The previous code that was returning a list of events that may or may
not be big, in any case, it required to keep track inside untangle of
what is the continuation when a file-descriptor is ready, duplicating
the work of epoll-events (that may also be big).

The new code should be better for the CPU and the RAM.